+++ databento/live/protocol.py
@@ -7,11 +7,9 @@ from functools import singledispatchmethod
 from typing import Final
 
 import databento_dbn
-from databento_dbn import DBNError
 from databento_dbn import Metadata
 from databento_dbn import Schema
 from databento_dbn import SType
-from databento_dbn import SystemCode
 from databento_dbn import VersionUpgradePolicy
 
 from databento.common import cram
@@ -313,13 +311,14 @@ class DatabentoLiveProtocol(asyncio.BufferedProtocol):
         list[SubscriptionRequest]
 
         """
-        logger.info(
-            "sending subscription to %s:%s %s start=%s snapshot=%s",
+        logger.debug(
+            "sending subscription request schema=%s stype_in=%s symbols='%s' start='%s' snapshot=%s id=%s",
             schema,
             stype_in,
             symbols,
             start if start is not None else "now",
             snapshot,
+            subscription_id,
         )
 
         stype_in_valid = validate_enum(stype_in, SType, "stype_in")
@@ -341,6 +340,12 @@ class DatabentoLiveProtocol(asyncio.BufferedProtocol):
             )
             subscriptions.append(message)
 
+        if len(subscriptions) > 1:
+            logger.debug(
+                "batched subscription into %d requests id=%s",
+                len(subscriptions),
+                subscription_id,
+            )
         self.transport.writelines(map(bytes, subscriptions))
         return subscriptions
 
@@ -374,7 +379,8 @@ class DatabentoLiveProtocol(asyncio.BufferedProtocol):
                     continue
                 if isinstance(record, databento_dbn.ErrorMsg):
                     logger.error(
-                        "gateway error: %s",
+                        "gateway error code=%s err='%s'",
+                        record.code,
                         record.err,
                     )
                     self._error_msgs.append(record.err)
@@ -382,19 +388,11 @@ class DatabentoLiveProtocol(asyncio.BufferedProtocol):
                     if record.is_heartbeat():
                         logger.debug("gateway heartbeat")
                     else:
-                        try:
-                            msg_code = record.code
-                        except DBNError:
-                            msg_code = None
-                        if msg_code == SystemCode.SLOW_READER_WARNING:
-                            logger.warning(
-                                record.msg,
-                            )
-                        else:
-                            logger.debug(
-                                "gateway message: %s",
-                                record.msg,
-                            )
+                        logger.info(
+                            "system message code=%s msg='%s'",
+                            record.code,
+                            record.msg,
+                        )
                 self.received_record(record)
 
     def _process_gateway(self, data: bytes) -> None:
@@ -423,11 +421,14 @@ class DatabentoLiveProtocol(asyncio.BufferedProtocol):
 
     @_handle_gateway_message.register(Greeting)
     def _(self, message: Greeting) -> None:
-        logger.debug("greeting received by remote gateway v%s", message.lsg_version)
+        logger.debug(
+            "greeting received by remote gateway version='%s'",
+            message.lsg_version,
+        )
 
     @_handle_gateway_message.register(ChallengeRequest)
     def _(self, message: ChallengeRequest) -> None:
-        logger.debug("received CRAM challenge: %s", message.cram)
+        logger.debug("received CRAM challenge cram='%s'", message.cram)
         response = cram.get_challenge_response(message.cram, self.__api_key)
         auth_request = AuthenticationRequest(
             auth=response,
@@ -435,22 +436,29 @@ class DatabentoLiveProtocol(asyncio.BufferedProtocol):
             ts_out=str(int(self._ts_out)),
             heartbeat_interval_s=self._heartbeat_interval_s,
         )
-        logger.debug("sending CRAM challenge response: %s", str(auth_request).strip())
+        logger.debug(
+            "sending CRAM challenge response auth='%s' dataset=%s encoding=%s ts_out=%s heartbeat_interval_s=%s client='%s'",
+            auth_request.auth,
+            auth_request.dataset,
+            auth_request.encoding,
+            auth_request.ts_out,
+            auth_request.heartbeat_interval_s,
+            auth_request.client,
+        )
         self.transport.write(bytes(auth_request))
 
     @_handle_gateway_message.register(AuthenticationResponse)
     def _(self, message: AuthenticationResponse) -> None:
         if message.success == "0":
-            logger.error("CRAM authentication failed: %s", message.error)
+            logger.error("CRAM authentication error: %s", message.error)
             self.authenticated.set_exception(
-                BentoError(f"User authentication failed: {message.error}"),
+                BentoError(message.error),
             )
             self.transport.close()
         else:
             session_id = message.session_id
 
             logger.debug(
-                "CRAM authenticated session id assigned `%s`",
-                session_id,
+                "CRAM authentication successful",
             )
             self.authenticated.set_result(session_id)