+++ examples/reference_security_master_get_range.py
@@ -1,6 +1,7 @@
 from pprint import pprint
 
 import pandas as pd
+
 from databento import Reference