+++ tests/test_historical_batch.py
@@ -4,12 +4,13 @@ from pathlib import Path
 from unittest.mock import MagicMock
 from zipfile import ZipFile
 
-import databento as db
 import pytest
 import requests
+from databento_dbn import Schema
+
+import databento as db
 from databento.common.publishers import Dataset
 from databento.historical.client import Historical
-from databento_dbn import Schema
 
 
 def test_batch_submit_job_given_invalid_schema_raises_error(
@@ -98,7 +99,7 @@ def test_batch_submit_job_sends_expected_request(
         "compression": "zstd",
         "pretty_px": False,
         "pretty_ts": False,
-        "map_symbols": False,
+        "map_symbols": True,
         "split_symbols": False,
         "split_duration": "day",
         "delivery": "download",