+++ tests/test_reference_corporate.py
@@ -4,13 +4,13 @@ from collections.abc import Iterable
 from pathlib import Path
 from unittest.mock import MagicMock
 
-import databento as db
 import pandas as pd
 import pytest
 import requests
 import zstandard
-from databento.reference.client import Reference
 
+import databento as db
+from databento.reference.client import Reference
 from tests import TESTS_ROOT