+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.67.0 - 2025-12-02
+
+#### Enhancements
+- Added a property `Live.subscription_requests` which returns a list of tuples containing every `SubscriptionRequest` for the live session
+- Changed the return value of `Live.subscribe()` to `int`, the value of the subscription ID, which can be used to index into the `Live.subscription_requests` property
+- Added feature to automatically monitor for hung connections in the `Live` client
+  - Hung connections will be disconnected client side with a `BentoError`
+- Added new venue, dataset, and publisher for Cboe Futures Exchange (`XCBF.PITCH`)
+
+#### Breaking changes
+- Several log messages have been reformatted to improve clarity and reduce redundancy, especially at debug levels
+- The `map_symbols` parameter for `Historical.batch.submit_job()` now defaults to `True` for JSON and CSV encodings
+
 ## 0.66.0 - 2025-11-18
 
 #### Enhancements